# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ConsoleRequestApplicant" : {
        "properties" : {
          "dui" : {
            "type" : "string"
          },
          "firstNames" : {
            "type" : "string"
          },
          "lastNames" : {
            "type" : "string"
          },
          "mobilePhone" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ConsoleRequestDetail" : {
        "properties" : {
          "applicant" : {
            "$ref" : "#/components/schemas/ConsoleRequestApplicant"
          },
          "completedSteps" : {
            "format" : "int32",
            "type" : "integer"
          },
          "expectedActivity" : {
            "$ref" : "#/components/schemas/ConsoleRequestExpectedActivity"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "identityDocument" : {
            "$ref" : "#/components/schemas/ConsoleRequestIdentityDocument"
          },
          "income" : {
            "$ref" : "#/components/schemas/ConsoleRequestIncome"
          },
          "number" : {
            "type" : "string"
          },
          "risk" : {
            "$ref" : "#/components/schemas/RiskAssessmentResponse"
          },
          "riskLevel" : {
            "enum" : [ "NOT_EVALUATED", "LOW", "PENDING_REVIEW", "MEDIUM", "HIGH" ],
            "type" : "string"
          },
          "signals" : {
            "$ref" : "#/components/schemas/ConsoleRequestSignals"
          },
          "startedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "IN_PROGRESS", "COMPLETED", "ABANDONED" ],
            "type" : "string"
          },
          "steps" : {
            "items" : {
              "$ref" : "#/components/schemas/ConsoleRequestStepTime"
            },
            "type" : "array"
          },
          "submittedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "timeline" : {
            "items" : {
              "$ref" : "#/components/schemas/ConsoleRequestTimelineEntry"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "ConsoleRequestDuiData" : {
        "properties" : {
          "birthDate" : {
            "format" : "date",
            "type" : "string"
          },
          "dui" : {
            "type" : "string"
          },
          "expiryDate" : {
            "format" : "date",
            "type" : "string"
          },
          "firstNames" : {
            "type" : "string"
          },
          "gender" : {
            "type" : "string"
          },
          "issueDate" : {
            "format" : "date",
            "type" : "string"
          },
          "lastNames" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ConsoleRequestExpectedActivity" : {
        "properties" : {
          "monthlyAmountRangeCode" : {
            "type" : "string"
          },
          "monthlyAmountRangeLabel" : {
            "type" : "string"
          },
          "registeredAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "transactionTypeCode" : {
            "type" : "string"
          },
          "transactionTypeLabel" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ConsoleRequestIdentityDocument" : {
        "properties" : {
          "attempts" : {
            "format" : "int32",
            "type" : "integer"
          },
          "checkDigitValid" : {
            "type" : "boolean"
          },
          "confidence" : {
            "type" : "number"
          },
          "confirmed" : {
            "$ref" : "#/components/schemas/ConsoleRequestDuiData"
          },
          "confirmedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "correctedFields" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "duiMatchesDeclared" : {
            "type" : "boolean"
          },
          "expired" : {
            "type" : "boolean"
          },
          "failure" : {
            "type" : "string"
          },
          "looksAuthentic" : {
            "type" : "boolean"
          },
          "namesMatchDeclared" : {
            "type" : "boolean"
          },
          "ocrModel" : {
            "type" : "string"
          },
          "ocrStatus" : {
            "enum" : [ "READ", "UNREADABLE", "FAILED" ],
            "type" : "string"
          },
          "photos" : {
            "items" : {
              "enum" : [ "FRONT", "BACK" ],
              "type" : "string"
            },
            "type" : "array"
          },
          "processedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "read" : {
            "$ref" : "#/components/schemas/ConsoleRequestDuiData"
          },
          "unreadableReason" : {
            "enum" : [ "BLURRY", "GLARE", "CROPPED", "TOO_DARK", "NOT_A_DUI", "WRONG_SIDES", "OTHER" ],
            "type" : "string"
          },
          "unreadableSide" : {
            "enum" : [ "FRONT", "BACK", "BOTH" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ConsoleRequestIncome" : {
        "properties" : {
          "rangeCode" : {
            "type" : "string"
          },
          "rangeLabel" : {
            "type" : "string"
          },
          "registeredAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "sourceCode" : {
            "type" : "string"
          },
          "sourceDetail" : {
            "type" : "string"
          },
          "sourceLabel" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ConsoleRequestIpDetails" : {
        "properties" : {
          "asName" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "country" : {
            "type" : "string"
          },
          "countryCode" : {
            "type" : "string"
          },
          "failure" : {
            "type" : "string"
          },
          "isp" : {
            "type" : "string"
          },
          "lookedUpAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "org" : {
            "type" : "string"
          },
          "region" : {
            "type" : "string"
          },
          "regionName" : {
            "type" : "string"
          },
          "timezone" : {
            "type" : "string"
          },
          "zip" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ConsoleRequestListItem" : {
        "properties" : {
          "completedSteps" : {
            "format" : "int32",
            "type" : "integer"
          },
          "date" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "monthlyAmountRangeLabel" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "number" : {
            "type" : "string"
          },
          "riskLevel" : {
            "enum" : [ "NOT_EVALUATED", "LOW", "PENDING_REVIEW", "MEDIUM", "HIGH" ],
            "type" : "string"
          },
          "status" : {
            "enum" : [ "IN_PROGRESS", "COMPLETED", "ABANDONED" ],
            "type" : "string"
          },
          "transactionTypeLabel" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ConsoleRequestSignals" : {
        "properties" : {
          "approximateLocation" : {
            "type" : "string"
          },
          "capturedAt" : {
            "description" : "Fecha y hora en que la app envió las señales por última vez",
            "format" : "date-time",
            "type" : "string"
          },
          "device" : {
            "type" : "string"
          },
          "deviceFingerprint" : {
            "type" : "string"
          },
          "ip" : {
            "type" : "string"
          },
          "ipDetails" : {
            "$ref" : "#/components/schemas/ConsoleRequestIpDetails"
          },
          "latitude" : {
            "type" : "number"
          },
          "locationStatus" : {
            "enum" : [ "AVAILABLE", "UNAVAILABLE" ],
            "type" : "string"
          },
          "longitude" : {
            "type" : "number"
          },
          "nightTime" : {
            "type" : "boolean"
          },
          "requestsFromSameDevice" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalDurationSeconds" : {
            "format" : "int64",
            "type" : "integer"
          },
          "typingPace" : {
            "enum" : [ "SLOW", "NORMAL", "FAST" ],
            "type" : "string"
          },
          "typingSpeedCpm" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "ConsoleRequestStepTime" : {
        "properties" : {
          "attempts" : {
            "format" : "int32",
            "type" : "integer"
          },
          "completedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "durationSeconds" : {
            "format" : "int32",
            "type" : "integer"
          },
          "startedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "step" : {
            "enum" : [ "PRIVACY_NOTICE", "BASIC_DATA", "IDENTITY_DOCUMENT", "INCOME", "EXPECTED_ACTIVITY", "REVIEW" ],
            "type" : "string"
          },
          "typingSpeedCps" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "ConsoleRequestTimelineEntry" : {
        "properties" : {
          "actor" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "occurredAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "type" : {
            "enum" : [ "REQUEST_STARTED", "PRIVACY_ACCEPTED", "BASIC_DATA_COMPLETED", "IDENTITY_DOCUMENT_CAPTURED", "IDENTITY_DOCUMENT_CONFIRMED", "INCOME_REGISTERED", "EXPECTED_ACTIVITY_REGISTERED", "REQUEST_SUBMITTED", "SUBMISSION_FAILED", "SCORE_ASSIGNED", "REQUEST_ABANDONED" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "OnboardingRequestResponse" : {
        "properties" : {
          "completedSteps" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "number" : {
            "type" : "string"
          },
          "riskLevel" : {
            "enum" : [ "NOT_EVALUATED", "LOW", "PENDING_REVIEW", "MEDIUM", "HIGH" ],
            "type" : "string"
          },
          "startedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "IN_PROGRESS", "COMPLETED", "ABANDONED" ],
            "type" : "string"
          },
          "submittedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PageResponseConsoleRequestListItem" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/ConsoleRequestListItem"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ProblemDetail" : {
        "description" : "Error de la API (RFC 9457)",
        "properties" : {
          "detail" : {
            "example" : "Onboarding request not found",
            "type" : "string"
          },
          "instance" : {
            "type" : "string"
          },
          "status" : {
            "example" : 404,
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "example" : "Not Found",
            "type" : "string"
          },
          "type" : {
            "example" : "about:blank",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PublishScoreRuleVersionRequest" : {
        "properties" : {
          "status" : {
            "description" : "DRAFT, PROVISIONAL o CONFIRMED",
            "enum" : [ "DRAFT", "PROVISIONAL", "CONFIRMED", "RETIRED" ],
            "example" : "CONFIRMED",
            "type" : "string"
          },
          "statusNote" : {
            "description" : "Motivo del cambio",
            "example" : "Umbral confirmado por Conozca a su Cliente",
            "maxLength" : 250,
            "minLength" : 0,
            "pattern" : "\\s*|[\\p{L}\\p{N}\\p{Sc} .,;:¿?¡!'\"‘’“”«»()/&%#@+*_°–—…\\-]*[\\p{L}\\p{N}][\\p{L}\\p{N}\\p{Sc} .,;:¿?¡!'\"‘’“”«»()/&%#@+*_°–—…\\-]*",
            "type" : "string"
          },
          "threshold" : {
            "description" : "Umbral mensual en USD",
            "example" : 600,
            "exclusiveMinimum" : 0,
            "type" : "number"
          }
        },
        "required" : [ "status", "threshold" ],
        "type" : "object"
      },
...
      "RiskAssessmentResponse" : {
        "properties" : {
          "evaluatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "evaluatedValue" : {
            "type" : "number"
          },
          "explanation" : {
            "type" : "string"
          },
          "level" : {
            "enum" : [ "NOT_EVALUATED", "LOW", "PENDING_REVIEW", "MEDIUM", "HIGH" ],
            "type" : "string"
          },
          "ruleCode" : {
            "type" : "string"
          },
          "ruleName" : {
            "type" : "string"
          },
          "ruleThreshold" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "ScoreRuleResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "evaluatedField" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "operator" : {
            "type" : "string"
          },
          "resultIfMatched" : {
            "enum" : [ "NOT_EVALUATED", "LOW", "PENDING_REVIEW", "MEDIUM", "HIGH" ],
            "type" : "string"
          },
          "resultIfNotMatched" : {
            "enum" : [ "NOT_EVALUATED", "LOW", "PENDING_REVIEW", "MEDIUM", "HIGH" ],
            "type" : "string"
          },
          "status" : {
            "enum" : [ "DRAFT", "PROVISIONAL", "CONFIRMED", "RETIRED" ],
            "type" : "string"
          },
          "statusNote" : {
            "type" : "string"
          },
          "threshold" : {
            "type" : "number"
          },
          "validFrom" : {
            "format" : "date-time",
            "type" : "string"
          },
          "validTo" : {
            "format" : "date-time",
            "type" : "string"
          },
          "version" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Token de POST /api/console/auth/login. Solo la consola lo usa; el flujo móvil es público.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "API del flujo de onboarding móvil (apertura de cuenta), del score de riesgo y de la consola administrativa.\n\n- Rutas `/api/onboarding/**` y `/api/catalogs`: públicas (app móvil).\n- Rutas `/api/console/**`: requieren `Authorization: Bearer <token>` de `POST /api/console/auth/login`.\n- Errores: `application/problem+json` con `status`, `title` y `detail`.\n- Los campos que aparecen como opcionales en los esquemas de respuesta pueden venir `null`.",
    "title" : "Verificación Online API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/console/admin/score-rules/{code}/versions" : {
      "get" : {
        "description" : "La vigente es la que no tiene `validTo` y no está en borrador.",
        "operationId" : "scoreRuleAdminVersions",
        "parameters" : [ {
          "in" : "path",
          "name" : "code",
          "required" : true,
          "schema" : {
            "maxLength" : 40,
            "minLength" : 0,
            "pattern" : "[A-Z0-9][A-Z0-9_-]*",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ScoreRuleResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Versiones, de la más reciente a la más antigua"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Falta el token o ya no es válido"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Solo administradores"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "La regla no existe"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Historial de versiones de una regla",
        "tags" : [ "Consola · Administración de reglas" ]
      },
      "post" : {
        "description" : "Copia la lógica de la última versión con el umbral y el estado nuevos. PROVISIONAL o CONFIRMED la aplican desde ahora y retiran la vigente; DRAFT solo la guarda. Las evaluaciones ya hechas conservan la versión con la que se calcularon.",
        "operationId" : "scoreRuleAdminPublish",
        "parameters" : [ {
          "in" : "path",
          "name" : "code",
          "required" : true,
          "schema" : {
            "maxLength" : 40,
            "minLength" : 0,
            "pattern" : "[A-Z0-9][A-Z0-9_-]*",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PublishScoreRuleVersionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ScoreRuleResponse"
                }
              }
            },
            "description" : "Versión publicada"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Datos inválidos o estado RETIRED"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Falta el token o ya no es válido"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Solo administradores"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "La regla no existe"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "La versión vigente empieza en el futuro"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Publica una versión nueva de una regla (por ejemplo, otro umbral)",
        "tags" : [ "Consola · Administración de reglas" ]
      }
    },
...
    "/api/console/requests" : {
      "get" : {
        "description" : "Filtra por estado, nivel de riesgo y texto (nombre o número). `page` empieza en 0; `size` máximo 100.",
        "operationId" : "consoleRequestList",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "IN_PROGRESS", "COMPLETED", "ABANDONED" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "riskLevel",
          "required" : false,
          "schema" : {
            "enum" : [ "NOT_EVALUATED", "LOW", "PENDING_REVIEW", "MEDIUM", "HIGH" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "maxLength" : 100,
            "minLength" : 0,
            "pattern" : "\\s*|[\\p{L}\\p{N}\\p{Sc} .,;:¿?¡!'\"‘’“”«»()/&%#@+*_°–—…\\-]*[\\p{L}\\p{N}][\\p{L}\\p{N}\\p{Sc} .,;:¿?¡!'\"‘’“”«»()/&%#@+*_°–—…\\-]*",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseConsoleRequestListItem"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Falta el token o ya no es válido"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Listado de solicitudes, de la más reciente a la más antigua",
        "tags" : [ "Consola · Solicitudes" ]
      }
    },
    "/api/console/requests/{requestId}" : {
      "get" : {
        "description" : "Datos declarados, ingresos, movimiento esperado, score y regla aplicada, señales, tiempo por paso y línea de tiempo.",
        "operationId" : "consoleRequestDetail",
        "parameters" : [ {
          "in" : "path",
          "name" : "requestId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsoleRequestDetail"
                }
              }
            },
            "description" : "Detalle de la solicitud"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Falta el token o ya no es válido"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "La solicitud no existe"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Detalle de una solicitud",
        "tags" : [ "Consola · Solicitudes" ]
      }
    },
...
    "/api/onboarding/requests/{requestId}" : {
      "get" : {
        "operationId" : "onboardingGet",
        "parameters" : [ {
          "in" : "path",
          "name" : "requestId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingRequestResponse"
                }
              }
            },
            "description" : "Resumen de la solicitud"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "La solicitud no existe"
          }
        },
        "summary" : "Consulta el resumen de una solicitud",
        "tags" : [ "Onboarding" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Dev",
    "url" : "https://api.dev.identidad.alambritos.online"
  }, {
    "description" : "QA",
    "url" : "https://api.qa.identidad.alambritos.online"
  }, {
    "description" : "Local",
    "url" : "http://localhost:8080"
  } ],
...
}